+++ front/src/app/app.config.ts
@@ -4,7 +4,7 @@ import { provideRouter, withViewTransitions } from '@angular/router';
 import { routes } from './app.routes';
 import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { provideIonicAngular } from '@ionic/angular/standalone';
-import { MiddlewareService } from './services/middleware.service';
+import { MiddlewareService } from './services/auth/middleware.service';
 
 export const appConfig: ApplicationConfig = {
   providers: [
