+++ front/src/app/services/api/region.service.ts
@@ -1,10 +1,10 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { map, Observable } from 'rxjs';
-import { Region } from '../model/db/Region';
-import { API_PATH } from './API';
-import { DbId } from '../model/db/dbDefs';
-import { RegionTree } from '../model/local/RegionTree';
+import { Region } from '../../model/db/Region';
+import { API_PATH } from '../API';
+import { DbId } from '../../model/db/dbDefs';
+import { RegionTree } from '../../model/local/RegionTree';
 
 @Injectable({
   providedIn: 'root'
